code checked and the basic version of the server is running - it's alive!

+++ Server/storage_handler.py
@@ -1,4 +1,3 @@
-
 import os
 import shutil
 import json
@@ -7,46 +6,63 @@ import base64
 from utils import create_path_string
 import consts
 
+
 # TODO: change
 def merge_results(project_id: str) -> dict:
-    results_path = create_path_string(consts.PROJECTS_DIRECTORY,
-                                      project_id,
+    results_path = create_path_string(consts.PROJECTS_DIRECTORY, project_id,
                                       consts.PROJECT_STORAGE_RESULTS)
 
     results = dict()
 
-    directories = (os.listdir(results_path)).sort()
+    directories = sorted(os.listdir(results_path))
 
     for directory in directories:
-        result_path = create_path_string(results_path, directory, 'results', consts.RESULT_FILE)
-        with open(result_path, 'r') as file:
+        result_file = create_path_string(results_path, directory,
+                                         consts.RETURNED_TASK_RESULTS_DIRECTORY,
+                                         consts.RESULTS_FILE, from_current_directory=False)
+        with open(result_file, 'r') as file:
             results.update(json.load(file))
-            os.remove(results_path)
+
+        os.remove(result_file)
+
+    results_json_file = create_path_string(results_path, consts.RESULTS_FILE,
+                                           from_current_directory=False)
+    with open(results_json_file, 'w') as file:
+        json.dump(results, file)
 
     return results
 
 
-
 def zip_additional_results(project_id) -> str:
+    base_results_path = create_path_string(consts.PROJECTS_DIRECTORY, project_id,
+                                           consts.PROJECT_STORAGE_RESULTS)
 
-    base_results_path = create_path_string(consts.PROJECTS_DIRECTORY,
-                                      project_id,
-                                      consts.PROJECT_STORAGE_RESULTS)
+    temp_results_path = create_path_string(base_results_path,
+                                           consts.TEMP_PROJECT_ADDITIONAL_RESULTS_DIRECTORY,
+                                           from_current_directory=False)
 
-    returned_results_path = create_path_string(base_results_path, consts.TEMP_PROJECT_ADDITIONAL_RESULTS_DIRECTORY)
+    os.makedirs(temp_results_path)
 
-    directories = sorted(os.listdir(base_results_path))
+    tasks_storage = sorted(os.listdir(base_results_path))
 
-    os.makedirs(returned_results_path)
+    for task_directory in tasks_storage:
+        if task_directory.isdigit():
+            result_path = create_path_string(base_results_path, task_directory,
+                                             consts.RETURNED_TASK_RESULTS_DIRECTORY,
+                                             consts.RETURNED_TASK_ADDITIONAL_RESULTS_DIRECTORY,
+                                             from_current_directory=False)
 
-    for directory in directories:
-        result_path = create_path_string(base_results_path, directory, 'results', consts.RETURNED_TASK_ADDITIONAL_RESULTS_DIRECTORY)
-        shutil.move(result_path, returned_results_path)
-        # TODO: maybe check if empty
-        os.rmdir(result_path)
+            additional_results = sorted(os.listdir(result_path))
+            for file in additional_results:
+                file_path = create_path_string(result_path, file, from_current_directory=False)
+                shutil.move(file_path, temp_results_path)
+            # TODO: maybe check if empty
+            os.rmdir(result_path)
 
-    zipped_results_path = shutil.make_archive(create_path_string(base_results_path, consts.RETURNED_TASK_ADDITIONAL_RESULTS_DIRECTORY),
-                        'zip', returned_results_path)
+    zipped_results_without_extension = create_path_string(base_results_path,
+                                                          consts.RETURNED_TASK_ADDITIONAL_RESULTS_DIRECTORY,
+                                                          from_current_directory=False)
+    zipped_results_path = shutil.make_archive(zipped_results_without_extension, 'zip', temp_results_path)
 
     with open(zipped_results_path, 'rb') as file:
         zipped_results = file.read()
@@ -54,6 +70,3 @@ def zip_additional_results(project_id) -> str:
     zipped_results = base64.b64encode(zipped_results).decode('utf-8')
 
     return zipped_results
-
-
-
